--- FOSSDB_web/apps/fossdb/hosting_platform/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class HostingPlatformConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "fossdb.hosting_platform"
