--- FOSSDB_web/apps/fossdb/license/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-
-from .models import License
-
-admin.site.register(License)
