--- FOSSDB_web/apps/fossdb/license/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class LicenseConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "fossdb.license"
