--- FOSSDB_web/apps/fossdb/programming_language/admin.py
@@ -1,11 +1,0 @@
-from django.contrib import admin
-
-from .models import ProgrammingLanguage, ProjectProgrammingLanguage
-
-
-class ProjectProgrammingLanguageInline(admin.TabularInline):
-    model = ProjectProgrammingLanguage
-    extra = 1
-
-
-admin.site.register(ProgrammingLanguage)
