--- FOSSDB_web/apps/fossdb/project/admin.py
@@ -1,20 +1,0 @@
-from django.contrib import admin
-
-from fossdb.hosting_platform.admin import ProjectHostingPlatformInline
-from fossdb.programming_language.admin import ProjectProgrammingLanguageInline
-
-from .models import Project
-
-
-class ProjectAdmin(admin.ModelAdmin):
-    inlines = [ProjectProgrammingLanguageInline, ProjectHostingPlatformInline]
-    list_display = ("author", "name", "_languages", "_hosting_platform")
-
-    def _languages(self, object):
-        return " | ".join([i.language.language for i in object.projectprogramminglanguage_set.all()])
-
-    def _hosting_platform(self, object):
-        return " | ".join([i.hosting_platform.hosting_platform for i in object.projecthostingplatform_set.all()])
-
-
-admin.site.register(Project, ProjectAdmin)
