--- FOSSDB_web/apps/fossdb/tag/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-
-from .models import Tag
-
-admin.site.register(Tag)
